pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vulnhuntr2"
version = "0.1.0"
description = "Mutation-driven smart contract security analysis with guard fragility scoring."
readme = "README.md"
license = "MIT"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.12,<3.13"

dependencies = [
    "rich>=13.7.1",
    "typer>=0.12.3",
    "python-dotenv==1.0.0",
    "tqdm>=4.66.0",
    "typing-extensions>=4.12.0",
    "urllib3>=2.2.0",
    "pydantic>=2.7.0",
    "tree-sitter>=0.21.0",
    "regex>=2024.5.15",
]

[project.optional-dependencies]
llm = ["openai>=1.35.0", "tenacity>=8.2.3"]
mutation = ["tree-sitter-languages>=1.10.2"]
ml = ["transformers>=4.41.0", "tokenizers>=0.19.1"]
static = ["slither-analyzer>=0.10.0"]
# mythril = ["mythril>=0.24.0"]  # Placeholder for future mythril integration
full = [
    "openai>=1.35.0",
    "tenacity>=8.2.3",
    "tree-sitter-languages>=1.10.2",
    "transformers>=4.41.0",
    "tokenizers>=0.19.1",
    "slither-analyzer>=0.10.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.7",
    "black>=24.4.2",
    "mypy>=1.10.0",
]

[project.scripts]
vulnhuntr = "vulnhuntr.cli:app"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["vulnhuntr*"]
exclude = ["tests*", "examples*", "vulnhunt*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
